fix: returning to static, relative path favicons
For some weird reason, when I used the absolute path, I was getting
mixed content errors in the browser console in dev and prod, as if
the favicon was being served from http on an https site. The absolute
path appeared to use https, so I couldn't tell what the issue was.
Instead of spinning my wheels on it, I'm just reverting back to the
relative path favicons which were working fine.